Bugfix: Django cannot handle mixed TestSuites
root cause is `django.test.runner.partition_suite` (*)
where it is assumed that the suite will only have nested suite's of
its own type. Thus a regular `unittest.suite.TestSuite` inside a
`MyTestSuite` will not be recognized, as it is not an instance of
`MyTestSuite`, and thus is attempted to be added as a test.

The different kinds come in probably at
`unittest.TestLoader.loadTestsFromTestCase` (through `defaultTestLoader`)
and thus `TestLoader.suiteClass` is still the vanilla `unittest` one

(*)

* https://github.com/django/django/blob/stable/1.8.x/django/test/runner.py#L308
* https://github.com/django/django/blob/stable/1.10.x/django/test/runner.py#L647
* introduced in
https://github.com/django/django/commit/4ba373840a7b299fff4a7bcc1001e32dffceee86